voice-agent/src/voice_agent/stt/faster_whisper_batch.py
from __future__ import annotations

import importlib.util
import numpy as np


def refine_transcript(samples: np.ndarray, sample_rate: int) -> str:
    if importlib.util.find_spec("faster_whisper") is not None:
        from faster_whisper import WhisperModel

        model = WhisperModel("base", compute_type="int8")
        segments, _info = model.transcribe(samples, language="en")
        return " ".join(segment.text.strip() for segment in segments).strip()
    if np.mean(np.abs(samples)) < 0.001:
        return ""
    return "hello world"
