voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Local Voice Authenticated Realtime Agent"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "websockets>=11.0",
  "pydantic>=2.6",
  "numpy>=1.24",
  "soundfile>=0.12",
  "pyyaml>=6.0",
  "requests>=2.31",
]

[project.optional-dependencies]
voice = ["sounddevice>=0.4"]

[project.scripts]
voice-agent = "voice_agent.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
